+++ src/diary/errors.rs
@@ -78,4 +78,7 @@ pub enum DiaryError {
 
     #[error("No content provided, aborting.")]
     NoContent,
+
+    #[error("Failed to init git repo.")]
+    GitError(#[from] git2::Error),
 }